feat: Reservation에 price 추가 (#108)
#99

+++ src/main/java/poomasi/domain/farm/service/FarmFarmerService.java
@@ -40,7 +40,7 @@ public class FarmFarmerService {
         return farmRepository.findByIdAndDeletedAtIsNull(farmId).orElseThrow(() -> new BusinessException(FARM_NOT_FOUND));
     }
 
-    public void updateFarmExpPrice(Long farmerId, Long farmId, Long expPrice) {
+    public void updateFarmExpPrice(Long farmerId, Long farmId, int expPrice) {
         Farm farm = this.getFarmByFarmId(farmId);
         if (!farm.getOwnerId().equals(farmerId)) {
             throw new BusinessException(FARM_OWNER_MISMATCH);
